- moved from float to double to avoid excessive casting from secondstring - added a few of the more powerful distances - fixed a bug in the VPTree builder (although is still not working as I expect it to)
git-svn-id: http://google-refine.googlecode.com/svn/trunk@248 7d457c2a-affb-35e4-300a-418c747d4874

+++ src/main/java/edu/mit/simile/vicino/Distance.java
@@ -2,6 +2,6 @@ package edu.mit.simile.vicino;
 
 public interface Distance {
 
-    public float d(String x, String y);
+    public double d(String x, String y);
 
 }
