+++ src/main/java/edu/mit/simile/vicino/distances/PPMDistance.java
@@ -8,7 +8,7 @@ import com.colloquial.arithcode.PPMModel;
 
 public class PPMDistance extends PseudoMetricDistance {
 
-    public float d2(String x, String y) {
+    public double d2(String x, String y) {
         String str = x + y;
         float result = 0.0f;
         try {
