+++ src/main/java/edu/mit/simile/vicino/distances/PseudoMetricDistance.java
@@ -4,14 +4,14 @@ import edu.mit.simile.vicino.Distance;
 
 public abstract class PseudoMetricDistance implements Distance {
 
-    public float d(String x, String y) {
-        float cxx = d2(x, x);
-        float cyy = d2(y, y);
-        float cxy = d2(x, y);
-        float cyx = d2(y, x);
-        float result1 = (cxy + cyx) / (cxx + cyy) - 1.0f;
+    public double d(String x, String y) {
+        double cxx = d2(x, x);
+        double cyy = d2(y, y);
+        double cxy = d2(x, y);
+        double cyx = d2(y, x);
+        double result1 = (cxy + cyx) / (cxx + cyy) - 1.0d;
         return result1;
     }
 
-    protected abstract float d2(String x, String y);
+    protected abstract double d2(String x, String y);
 }
